fix package issue on js directory

+++ Makefile
@@ -38,7 +38,7 @@ package: rentroll
 	cp rentroll ./tmp/rentroll/
 	cp conf.json ./tmp/rentroll/
 	cp -r html ./tmp/rentroll/
-	cp -r js ./tmp/rentroll/
+	if [ -e js ]; then cp -r js ./tmp/rentroll/ ; fi
 	cp activate.sh update.sh ./tmp/rentroll/
 	rm -f ./rrnewdb ./rrbkup ./rrrestore
 	ln -s tmp/rentroll/rrnewdb
